refactor: 댓글 입력기능 리팩토링
