+++ src/app/components/common/comment/CommentInput.tsx
@@ -1,44 +1,26 @@
 'use client';
 
-import styles from './comment.module.scss'
-import { KeyboardEvent, useContext } from 'react';
+import styles from './comment.module.scss';
+import { KeyboardEvent } from 'react';
 import useInput from '@/app/lib/hooks/useInput';
-import { CommentListContext } from '@/app/components/common/comment/CommentListProvider';
+import { postRecruitmentBoardComment } from '@/app/lib/apis/recruitment-boards/recruitmentBoard';
+import useCommentInput from '@/app/lib/hooks/useCommentInput';
 
-export default function CommentInput() {
-  const [content, onChange, reset] = useInput({ comment: '' });
-  const { boardId, commentTargetBoardType } = useContext(CommentListContext);
+export interface Props {
+  boardId: string;
+}
 
-  const onSubmit = (e: KeyboardEvent<HTMLInputElement>) => {
-    if (!content.comment.trim()) {
-      return;
-    }
-
-    if (e.key === 'Enter') {
-      const API_URL = process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3000';
-      const token = process.env.NEXT_PUBLIC_ACCESS_TOKEN;
-      console.log(content.comment);
-      try {
-        const response = fetch(`${API_URL}/api/v1/comments/${boardId}?commentTargetBoardType=${commentTargetBoardType}`, {
-          method: 'POST',
-          redirect: 'follow',
-          headers: { Authorization: `Bearer ${token}`, 'Content-Type': 'application/json' },
-          body: JSON.stringify({
-            content: content.comment,
-            groupId: null,
-          }),
-        });
-      } catch (error) {
-        console.error(error);
-      }
-
-      reset();
-      window.location.reload();
-    }
-  }
+export default function CommentInput({ boardId }: Props) {
+  const { content, onChange, onSubmit } = useCommentInput(boardId);
 
   return (
-    <input className={styles.input} placeholder={'댓글 추가'} name={'comment'} value={content.comment} onChange={onChange}
-           onKeyDown={onSubmit}/>
-  )
+    <input
+      className={styles.input}
+      placeholder={'댓글 추가'}
+      name={'comment'}
+      value={content.comment}
+      onChange={onChange}
+      onKeyDown={onSubmit}
+    />
+  );
 }
